+++ addons/yat/src/classes/managers/command_manager/CommandManager.cs
@@ -9,7 +9,6 @@ using YAT.Enums;
 using YAT.Helpers;
 using YAT.Interfaces;
 using YAT.Scenes;
-using YAT.Scenes.BaseTerminal;
 using YAT.Types;
 
 namespace YAT.Classes.Managers;