+++ addons/yat/src/commands/Stats.cs
@@ -4,8 +4,7 @@ using Godot;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
-using YAT.Scenes.BaseTerminal;
-using YAT.Scenes.Monitor;
+using YAT.Scenes;
 using YAT.Types;
 
 namespace YAT.Commands;