+++ addons/yat/src/commands/View.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
-using YAT.Scenes.BaseTerminal;
+using YAT.Scenes;
 using YAT.Types;
 using static Godot.RenderingServer;