+++ addons/yat/src/commands/Wenv.cs
@@ -2,7 +2,7 @@ using Godot;
 using YAT.Attributes;
 using YAT.Enums;
 using YAT.Interfaces;
-using YAT.Scenes.BaseTerminal;
+using YAT.Scenes;
 using YAT.Types;
 
 namespace YAT.Commands;