+++ addons/yat/src/scenes/base_terminal/components/input_info/InputInfo.cs
@@ -1,6 +1,6 @@
 using Godot;
 
-namespace YAT.Scenes.BaseTerminal.Components.InputInfo;
+namespace YAT.Scenes;
 
 public partial class InputInfo : PanelContainer
 {