+++ addons/yat/src/scenes/game_terminal/components/quick_commands_context/QuickCommandsContext.cs
@@ -1,8 +1,6 @@
 using YAT.Helpers;
-using YAT.Scenes.GameTerminal.Components;
-using YAT.Scenes.YatWindow;
 
-namespace YAT.Scenes.BaseTerminal;
+namespace YAT.Scenes;
 
 public partial class QuickCommandsContext : ContextSubmenu
 {